Refactored, added payment controller
