+++ server/.env
@@ -7,3 +7,5 @@ ATLAS_URI="mongodb+srv://team-lavender:lavender-team@cluster0.kenuz.mongodb.net/
 client_id=  "871373961261-rjej65g97dc3o6jiuflq6s2gp5v9ptut.apps.googleusercontent.com"
 client_secret=  "XauLtqJv2LbkiRd7nTMhRnQq"
 redirect_uris=  "http://localhost:3000"
+
+STRIPE_SK = "sk_test_51HJRdeB2HY6qlBJzsa1OP1JDhvYohU5K7gSEhjKSLCjOZlYL9s7kuzJ1AdOMGzG6LkqZhQHTW1rcJ8D8iktzmw5k00ONGnd4RX"
